+++ src/Fake.Dotnet/Dotnet.fs
@@ -3,31 +3,30 @@ module Fake.Dotnet
 
 open Fake
 open FSharp.Data
-open FSharp.Data.JsonExtensions
 open System
 open System.IO
 open System.Security.Cryptography
 open System.Text
 
 /// .NET Core SDK default install directory (set to default localappdata dotnet dir). Update this to redirect all tool commands to different location. 
-let mutable DefaultDotnetCliDir = environVar "LocalAppData" @@ "Microsoft" @@ "dotnet"
+let DefaultDotnetSdkDir = environVar "LocalAppData" @@ "Microsoft" @@ "dotnet"
 
 /// Get dotnet cli executable path
 /// ## Parameters
 ///
-/// - 'dotnetCliDir' - dotnet cli install directory 
-let private dotnetCliPath dotnetCliDir = dotnetCliDir @@ "dotnet.exe"
+/// - 'dotnetSdkDir' - dotnet cli install directory 
+let private dotnetCliToolPath dotnetSdkDir = dotnetSdkDir @@ "dotnet.exe"
 
 /// Get .NET Core SDK download uri
-let private getDotnetCliInstallerUrl branch = sprintf "https://raw.githubusercontent.com/dotnet/cli/%s/scripts/obtain/dotnet-install.ps1" branch
+let private getDotnetSdkInstallerUrl branch = sprintf "https://raw.githubusercontent.com/dotnet/cli/%s/scripts/obtain/dotnet-install.ps1" branch
 
 /// Download .NET Core SDK installer
-let private downloadDotnetInstaller branch fileName =  
-    let url = getDotnetCliInstallerUrl branch
+let private downloadDotnetSdkInstaller branch fileName =  
+    let url = getDotnetSdkInstallerUrl branch
     let installScript = Http.RequestStream url
     use outFile = File.Open(fileName, FileMode.Create)
     installScript.ResponseStream.CopyTo(outFile)
-    trace (sprintf "downloaded dotnet installer (%s) to %s" url fileName)
+    trace (sprintf "downloaded dotnet sdk installer (%s) to %s" url fileName)
 
 /// [omit]
 let private md5 (data : byte array) : string =
@@ -38,7 +37,7 @@ let private md5 (data : byte array) : string =
 
 
 /// .NET Core SDK installer download options
-type DotNetInstallerOptions =
+type DotnetSdkInstallerOptions =
     {   
         /// Always download install script (otherwise install script is cached in temporary folder)
         AlwaysDownload: bool;
@@ -56,29 +55,29 @@ type DotNetInstallerOptions =
 /// ## Parameters
 ///
 /// - 'setParams' - set download installer options
-let DotnetDownloadInstaller setParams =
-    let param = DotNetInstallerOptions.Default |> setParams
+let DotnetSdkDownloadInstaller setParams =
+    let param = DotnetSdkInstallerOptions.Default |> setParams
 
     let scriptName = sprintf "dotnet_install_%s.ps1" <| md5 (Encoding.ASCII.GetBytes(param.Branch))
     let tempInstallerScript = Path.GetTempPath() @@ scriptName
 
     // maybe download installer script
     match param.AlwaysDownload || not(File.Exists(tempInstallerScript)) with
-        | true -> downloadDotnetInstaller param.Branch tempInstallerScript 
+        | true -> downloadDotnetSdkInstaller param.Branch tempInstallerScript 
         | _ -> ()
 
     tempInstallerScript
 
 
 /// .NET Core SDK architecture
-type DotnetCliArchitecture =
+type DotnetSdkArchitecture =
     /// this value represents currently running OS architecture 
     | Auto
     | X86
     | X64
 
 /// .NET Core SDK version (used to specify version when installing .NET Core SDK)
-type DotnetCliVersion =
+type DotnetSdkVersion =
     /// most latest build on specific channel 
     | Latest
     ///  last known good version on specific channel (Note: LKG work is in progress. Once the work is finished, this will become new default)
@@ -87,56 +86,61 @@ type DotnetCliVersion =
     | Version of string
   
 /// .NET Core SDK install options
-type DotNetCliInstallOptions =
+type DotnetSdkInstallOptions =
     {   
         /// Custom installer obtain (download) options
-        InstallerOptions: DotNetInstallerOptions -> DotNetInstallerOptions
+        InstallerOptions: DotnetSdkInstallerOptions -> DotnetSdkInstallerOptions
         /// .NET Core SDK channel (defaults to normalized installer branch)
         Channel: string option;
         /// .NET Core SDK version
-        Version: DotnetCliVersion;
+        Version: DotnetSdkVersion;
         /// Custom installation directory (for local build installation)
         CustomInstallDir: string option
         /// Architecture
-        Architecture: DotnetCliArchitecture;
+        Architecture: DotnetSdkArchitecture;
+        /// Installs just the shared runtime bits, not the entire SDK
+        SharedRuntime: bool;
         /// Include symbols in the installation (Switch does not work yet. Symbols zip is not being uploaded yet) 
         DebugSymbols: bool;
         /// If set it will not perform installation but instead display what command line to use
         DryRun: bool
         /// Do not update path variable
         NoPath: bool
+        /// Displays diagnostics information.
+        Verbose: bool
+        ///  If set, the installer will use the proxy when making web requests
+        ProxyAddress: string option
     }
 
     /// Parameter default values.
     static member Default = {
         InstallerOptions = id
         Channel = None
-        Version = Latest        
+        Version = Latest
         CustomInstallDir = None
-        Architecture = Auto        
+        Architecture = Auto
+        SharedRuntime = false
         DebugSymbols = false
         DryRun = false
         NoPath = true
+        Verbose = false
+        ProxyAddress = None
     }
 
-// Preview2 options for concrete version
-let Preview2ToolingVersionOptions options (version: string) = 
-    { options with
-        InstallerOptions = (fun io -> 
-            { io with
-                Branch = "v1.0.0-preview2"                    
-            })
-        Channel = Some "preview"
-        Version = Version (sprintf "1.0.0-preview2-%s" version)
-    }
+/// Well known DotnetSdk versions
+type SdkVersions =
 
-/// .NET Core SDK install options preconfigured for preview2 tooling (1.0.0 .NET core version)
-let Preview2ToolingOptions options = 
-    Preview2ToolingVersionOptions options "003121"
+    /// Version shipped with VS2017 (1.0.0 SDK, shared framework 1.0.4 and 1.1.1)
+    static member NetCore100 options: DotnetSdkInstallOptions = 
+        { options with
+            Version = Version "1.0.0"
+        }
 
-/// .NET Core SDK install options preconfigured for preview2 tooling (1.0.1 .NET core version)
-let Preview2Tooling101Options options = 
-    Preview2ToolingVersionOptions options "003131"
+    /// Released 2017/03/07 (1.0.1 SDK, shared framework 1.0.4 and 1.1.1)
+    static member NetCore101 options: DotnetSdkInstallOptions = 
+        { options with
+            Version = Version "1.0.1"
+        }
 
 /// [omit]
 let private optionToParam option paramFormat =
@@ -151,7 +155,7 @@ let private boolToFlag value flagParam =
     | false -> ""
 
 /// [omit]
-let private buildDotnetCliInstallArgs (param: DotNetCliInstallOptions) =
+let private buildDotnetSdkInstallArgs (param: DotnetSdkInstallOptions) =
     let versionParamValue = 
         match param.Version with
         | Latest -> "latest"
@@ -163,7 +167,7 @@ let private buildDotnetCliInstallArgs (param: DotNetCliInstallOptions) =
         match param.Channel with
             | Some ch -> ch
             | None -> 
-                let installerOptions = DotNetInstallerOptions.Default |> param.InstallerOptions
+                let installerOptions = DotnetSdkInstallerOptions.Default |> param.InstallerOptions
                 installerOptions.Branch |> replace "/" "-"
 
     let architectureParamValue = 
@@ -176,9 +180,12 @@ let private buildDotnetCliInstallArgs (param: DotNetCliInstallOptions) =
         sprintf "-Version '%s'" versionParamValue        
         optionToParam architectureParamValue "-Architecture %s"
         optionToParam param.CustomInstallDir "-InstallDir '%s'"
+        optionToParam param.ProxyAddress "-ProxyAddress '%s'"
+        boolToFlag param.SharedRuntime "-SharedRuntime"
         boolToFlag param.DebugSymbols "-DebugSymbols"
         boolToFlag param.DryRun "-DryRun"
         boolToFlag param.NoPath "-NoPath"
+        boolToFlag param.Verbose "-Verbose"        
     ] |> Seq.filter (not << String.IsNullOrEmpty) |> String.concat " "
 
 
@@ -187,11 +194,11 @@ let private buildDotnetCliInstallArgs (param: DotNetCliInstallOptions) =
 /// ## Parameters
 ///
 /// - 'setParams' - set installation options
-let DotnetCliInstall setParams =
-    let param = DotNetCliInstallOptions.Default |> setParams  
-    let installScript = DotnetDownloadInstaller param.InstallerOptions
+let DotnetSdkInstall setParams =
+    let param = DotnetSdkInstallOptions.Default |> setParams  
+    let installScript = DotnetSdkDownloadInstaller param.InstallerOptions
 
-    let args = sprintf "-ExecutionPolicy Bypass -NoProfile -NoLogo -NonInteractive -Command \"%s %s; if (-not $?) { exit -1 };\"" installScript (buildDotnetCliInstallArgs param)
+    let args = sprintf "-ExecutionPolicy Bypass -NoProfile -NoLogo -NonInteractive -Command \"%s %s; if (-not $?) { exit -1 };\"" installScript (buildDotnetSdkInstallArgs param)
     let exitCode = 
         ExecProcess (fun info ->
             info.FileName <- "powershell"
@@ -202,7 +209,7 @@ let DotnetCliInstall setParams =
     if exitCode <> 0 then
         // force download new installer script
         traceError ".NET Core SDK install failed, trying to redownload installer..."
-        DotnetDownloadInstaller (param.InstallerOptions >> (fun o -> 
+        DotnetSdkDownloadInstaller (param.InstallerOptions >> (fun o -> 
             { o with 
                 AlwaysDownload = true
             })) |> ignore
@@ -211,8 +218,8 @@ let DotnetCliInstall setParams =
 /// dotnet cli command execution options
 type DotnetOptions =
     {
-        /// Dotnet cli executable path
-        DotnetCliPath: string;
+        /// Dotnet sdk install directory
+        DotnetSdkDir: string;
         /// Command working directory
         WorkingDirectory: string;
         /// Custom parameters
@@ -220,7 +227,7 @@ type DotnetOptions =
     }
 
     static member Default = {
-        DotnetCliPath = dotnetCliPath DefaultDotnetCliDir
+        DotnetSdkDir = DefaultDotnetSdkDir
         WorkingDirectory = currentDirectory
         CustomParams = None
     }
@@ -229,9 +236,11 @@ type DotnetOptions =
 /// Execute raw dotnet cli command
 /// ## Parameters
 ///
-/// - 'options' - common execution options
+/// - 'setOptions' - set dotnet execution options
 /// - 'args' - command arguments
-let Dotnet (options: DotnetOptions) args = 
+let Dotnet (setOptions: DotnetOptions -> DotnetOptions) args = 
+    let options = DotnetOptions.Default |> setOptions    
+
     let errors = new System.Collections.Generic.List<string>()
     let messages = new System.Collections.Generic.List<string>()
     let timeout = TimeSpan.MaxValue
@@ -250,7 +259,7 @@ let Dotnet (options: DotnetOptions) args =
 
     let result = 
         ExecProcessWithLambdas (fun info ->
-            info.FileName <- options.DotnetCliPath
+            info.FileName <- dotnetCliToolPath options.DotnetSdkDir
             info.WorkingDirectory <- options.WorkingDirectory
             info.Arguments <- cmdArgs
         ) timeout true errorF messageF
@@ -271,21 +280,22 @@ let private argOption name value =
         | false -> ""
 
 /// dotnet restore verbosity
-type NugetRestoreVerbosity =
-    | Debug
-    | Verbose
-    | Information
+type DotnetVerbosity =
+    | Quiet
     | Minimal
-    | Warning
-    | Error
+    | Normal
+    | Detailed
+    | Diagnostic
 
 /// dotnet restore command options
 type DotnetRestoreOptions =
     {   
         /// Common tool options
-        Common: DotnetOptions;
-        /// Nuget feeds to search updates in. Use default if empty.
-        Sources: string list;
+        CommonOptions: DotnetOptions -> DotnetOptions;
+        /// Specifies a NuGet package source to use during the restore (-s|--source).
+        Source: string option;
+        /// Target runtime to restore packages for. (-r|--runtime <RUNTIME_IDENTIFIER>).
+        Runtime: string option;
         /// Directory to install packages in (--packages).
         Packages: string list;
         /// Path to the nuget configuration file (nuget.config).
@@ -293,38 +303,44 @@ type DotnetRestoreOptions =
         /// No cache flag (--no-cache)
         NoCache: bool;
         /// Restore logging verbosity (--verbosity)
-        Verbosity: NugetRestoreVerbosity option
+        Verbosity: DotnetVerbosity option
         /// Only warning failed sources if there are packages meeting version requirement (--ignore-failed-sources)
         IgnoreFailedSources: bool;
         /// Disables restoring multiple projects in parallel (--disable-parallel)
         DisableParallel: bool;
+        /// Set this flag to ignore project to project references and only restore the root project (--no-dependencies)
+        NoDependencies: bool;
     }
 
     /// Parameter default values.
     static member Default = {
-        Common = DotnetOptions.Default
-        Sources = []
+        CommonOptions = id
+        Source = None
+        Runtime = None
         Packages = []
         ConfigFile = None        
         NoCache = false
         Verbosity = None
         IgnoreFailedSources = false
         DisableParallel = false
+        NoDependencies = false
     }
 
 /// [omit]
 let private buildRestoreArgs (param: DotnetRestoreOptions) =
-    [   param.Sources |> argList2 "source"
+    [   param.Source |> Option.toList |> argList2 "source"
+        param.Runtime |> Option.toList |> argList2 "runtime"
         param.Packages |> argList2 "packages"
         param.ConfigFile |> Option.toList |> argList2 "configFile"
         param.NoCache |> argOption "no-cache" 
         param.IgnoreFailedSources |> argOption "ignore-failed-sources" 
         param.DisableParallel |> argOption "disable-parallel" 
+        param.NoDependencies |> argOption "no-dependencies" 
         param.Verbosity |> Option.toList |> Seq.map (fun v -> v.ToString()) |> argList2 "verbosity"
     ] |> Seq.filter (not << String.IsNullOrEmpty) |> String.concat " "
 
 
-/// Execute dotnet restore command
+/// Execute dotnet restore (Restore dependencies specified in the .NET project)
 /// ## Parameters
 ///
 /// - 'setParams' - set restore command parameters
@@ -333,7 +349,7 @@ let DotnetRestore setParams project =
     traceStartTask "Dotnet:restore" project
     let param = DotnetRestoreOptions.Default |> setParams    
     let args = sprintf "restore %s %s" project (buildRestoreArgs param)
-    let result = Dotnet param.Common args    
+    let result = Dotnet param.CommonOptions args    
     if not result.OK then failwithf "dotnet restore failed with code %i" result.ExitCode
     traceEndTask "Dotnet:restore" project
 
@@ -355,10 +371,10 @@ let private buildConfigurationArg (param: BuildConfiguration) =
 type DotNetPackOptions =
     {   
         /// Common tool options
-        Common: DotnetOptions;
+        CommonOptions: DotnetOptions -> DotnetOptions;
         /// Pack configuration (--configuration)
         Configuration: BuildConfiguration;
-        /// Version suffix to use
+        /// Defines the value for the $(VersionSuffix) property in the project
         VersionSuffix: string option;
         /// Build base path (--build-base-path)
         BuildBasePath: string option;
@@ -366,16 +382,28 @@ type DotNetPackOptions =
         OutputPath: string option;
         /// No build flag (--no-build)
         NoBuild: bool;
+        /// Include packages with symbols in addition to regular packages in output directory (--include-symbols)
+        IncludeSymbols: bool;
+        /// Include PDBs and source files. Source files go into the src folder in the resulting nuget package (--include-source)
+        IncludeSource: bool;
+        /// Set the serviceable flag in the package. For more information, please see https://aka.ms/nupkgservicing (-s|--serviceable)
+        Serviceable: bool;
+        /// Set the verbosity level of the command. (--verbosity)
+        Verbosity: DotnetVerbosity option
     }
 
     /// Parameter default values.
     static member Default = {
-        Common = DotnetOptions.Default
+        CommonOptions = id
         Configuration = Release
         VersionSuffix = None
         BuildBasePath = None
         OutputPath = None
         NoBuild = false
+        IncludeSymbols = false
+        IncludeSource = false
+        Serviceable = false
+        Verbosity = None
     }
 
 /// [omit]
@@ -386,6 +414,10 @@ let private buildPackArgs (param: DotNetPackOptions) =
         param.BuildBasePath |> Option.toList |> argList2 "build-base-path"
         param.OutputPath |> Option.toList |> argList2 "output"
         param.NoBuild |> argOption "no-build" 
+        param.IncludeSymbols |> argOption "include-symbols" 
+        param.IncludeSource |> argOption "include-source" 
+        param.Serviceable |> argOption "serviceable" 
+        param.Verbosity |> Option.toList |> Seq.map (fun v -> v.ToString()) |> argList2 "verbosity"
     ] |> Seq.filter (not << String.IsNullOrEmpty) |> String.concat " "
 
 
@@ -398,7 +430,7 @@ let DotnetPack setParams project =
     traceStartTask "Dotnet:pack" project
     let param = DotNetPackOptions.Default |> setParams    
     let args = sprintf "pack %s %s" project (buildPackArgs param)
-    let result = Dotnet param.Common args    
+    let result = Dotnet param.CommonOptions args    
     if not result.OK then failwithf "dotnet pack failed with code %i" result.ExitCode
     traceEndTask "Dotnet:pack" project
 
@@ -407,33 +439,30 @@ let DotnetPack setParams project =
 type DotNetPublishOptions =
     {   
         /// Common tool options
-        Common: DotnetOptions;
+        CommonOptions: DotnetOptions -> DotnetOptions;
         /// Pack configuration (--configuration)
         Configuration: BuildConfiguration;
         /// Target framework to compile for (--framework)
         Framework: string option;
         /// Target runtime to publish for (--runtime)
         Runtime: string option;
-        /// Build base path (--build-base-path)
-        BuildBasePath: string option;
         /// Output path (--output)
         OutputPath: string option;
-        /// Defines what `*` should be replaced with in version field in project.json (--version-suffix)
+        /// Defines the value for the $(VersionSuffix) property in the project(--version-suffix)
         VersionSuffix: string option;
-        /// No build flag (--no-build)
-        NoBuild: bool;
+        /// Set the verbosity level of the command. (--verbosity)
+        Verbosity: DotnetVerbosity option
     }
 
     /// Parameter default values.
     static member Default = {
-        Common = DotnetOptions.Default
+        CommonOptions = id
         Configuration = Release
         Framework = None
         Runtime = None
-        BuildBasePath = None
         OutputPath = None
         VersionSuffix = None
-        NoBuild = false
+        Verbosity = None
     }
 
 /// [omit]
@@ -442,10 +471,9 @@ let private buildPublishArgs (param: DotNetPublishOptions) =
         buildConfigurationArg param.Configuration
         param.Framework |> Option.toList |> argList2 "framework"
         param.Runtime |> Option.toList |> argList2 "runtime"
-        param.BuildBasePath |> Option.toList |> argList2 "build-base-path"
         param.OutputPath |> Option.toList |> argList2 "output"
         param.VersionSuffix |> Option.toList |> argList2 "version-suffix"
-        param.NoBuild |> argOption "no-build" 
+        param.Verbosity |> Option.toList |> Seq.map (fun v -> v.ToString()) |> argList2 "verbosity"
     ] |> Seq.filter (not << String.IsNullOrEmpty) |> String.concat " "
 
 
@@ -458,7 +486,7 @@ let DotnetPublish setParams project =
     traceStartTask "Dotnet:publish" project
     let param = DotNetPublishOptions.Default |> setParams    
     let args = sprintf "publish %s %s" project (buildPublishArgs param)
-    let result = Dotnet param.Common args    
+    let result = Dotnet param.CommonOptions args    
     if not result.OK then failwithf "dotnet publish failed with code %i" result.ExitCode
     traceEndTask "Dotnet:publish" project
 
@@ -467,30 +495,36 @@ let DotnetPublish setParams project =
 type DotNetBuildOptions =
     {   
         /// Common tool options
-        Common: DotnetOptions;
+        CommonOptions: DotnetOptions -> DotnetOptions;
         /// Pack configuration (--configuration)
         Configuration: BuildConfiguration;
         /// Target framework to compile for (--framework)
         Framework: string option;
         /// Target runtime to publish for (--runtime)
         Runtime: string option;
-        /// Build base path (--build-base-path)
-        BuildBasePath: string option;
         /// Output path (--output)
         OutputPath: string option;
-        /// Native flag (--native)
-        Native: bool;
+        /// Disables incremental build (--no-incremental)
+        NoIncremental: bool;
+        /// Set this flag to ignore project-to-project references and only build the root project (--no-dependencies)
+        NoDependencies: bool;
+        /// Defines the value for the $(VersionSuffix) property in the project(--version-suffix)
+        VersionSuffix: string option;
+        /// Set the verbosity level of the command. (--verbosity)
+        Verbosity: DotnetVerbosity option
     }
 
     /// Parameter default values.
     static member Default = {
-        Common = DotnetOptions.Default
+        CommonOptions = id
         Configuration = Release
         Framework = None
         Runtime = None
-        BuildBasePath = None
         OutputPath = None
-        Native = false
+        NoIncremental = false
+        NoDependencies = false
+        VersionSuffix = None
+        Verbosity = None
     }
 
 
@@ -500,9 +534,11 @@ let private buildBuildArgs (param: DotNetBuildOptions) =
         buildConfigurationArg param.Configuration
         param.Framework |> Option.toList |> argList2 "framework"
         param.Runtime |> Option.toList |> argList2 "runtime"
-        param.BuildBasePath |> Option.toList |> argList2 "build-base-path"
         param.OutputPath |> Option.toList |> argList2 "output"
-        (if param.Native then "--native" else "")
+        param.NoIncremental |> argOption "no-incremental" 
+        param.NoDependencies |> argOption "no-dependencies" 
+        param.VersionSuffix |> Option.toList |> argList2 "version-suffix"
+        param.Verbosity |> Option.toList |> Seq.map (fun v -> v.ToString()) |> argList2 "verbosity"
     ] |> Seq.filter (not << String.IsNullOrEmpty) |> String.concat " "
 
 
@@ -510,20 +546,27 @@ let private buildBuildArgs (param: DotNetBuildOptions) =
 /// ## Parameters
 ///
 /// - 'setParams' - set compile command parameters
-/// - 'project' - project to compile
-let DotnetCompile setParams project =    
-    traceStartTask "Dotnet:build" project
+/// - 'projectFile' - projectFile to build
+let DotnetBuild setParams projectFile =    
+    traceStartTask "Dotnet:build" projectFile
     let param = DotNetBuildOptions.Default |> setParams    
-    let args = sprintf "build %s %s" project (buildBuildArgs param)
-    let result = Dotnet param.Common args    
+    let args = sprintf "build %s %s" projectFile (buildBuildArgs param)
+    let result = Dotnet param.CommonOptions args    
     if not result.OK then failwithf "dotnet build failed with code %i" result.ExitCode
-    traceEndTask "Dotnet:build" project
+    traceEndTask "Dotnet:build" projectFile
 
-/// get sdk version from global.json
+
+/// Execute dotnet msbuild command using FAKE msbuild options
 /// ## Parameters
 ///
-/// - 'project' - global.json path
-let GlobalJsonSdk project =
-    let data = ReadFileAsString project
-    let info = JsonValue.Parse(data)
-    info?sdk?version.AsString()   
+/// - 'setMsbuildParams' - set msbuild command parameters
+/// - 'setOptions' - set dotnet execution options
+/// - 'project' - project to compile
+let DotnetMsbuild setMsbuildParams setOptions projectFile = 
+    traceStartTask "Dotnet:msbuild" projectFile
+
+    // build args
+    let args = MSBuildDefaults |> setMsbuildParams |> serializeMSBuildParams
+    let result = Dotnet setOptions (sprintf "msbuild %s %s" projectFile args) 
+    if not result.OK then failwithf "dotnet msbuild failed with code %i" result.ExitCode
+    traceEndTask "Dotnet:msbuild" projectFile