+++ release_notes.md
@@ -1,3 +1,14 @@
+#### 2.0.0
+* dropped support for DNX projects
+* removed Preview2ToolingOptions and Preview2Tooling101Options
+* removed GlobalJsonSdk helper
+* added shortcut versions helpers for 1.0.0 and 1.0.1 .NET Core SDKs (NetCore100SdkOptions, NetCore101SdkOptions)
+* renamed DotnetCompile to DotnetBuild
+* updated Dotnet* helpers to match current rtw cli options (now based on msbuild)
+* multiple renames to rebrand some DotnetCli helpers to DotnetSdk
+* make CommonOptions function
+* added DotnetMsbuild helper using FAKE msbuild options delegating to dotnet msbuild cli command
+
 #### 1.1.1
 * add preview2 tooling .NET Core 1.0.1 shortcut (Preview2Tooling101Options)